+++ R/plotImputation.R
@@ -9,6 +9,8 @@
 #' @param assayImputation Character scalar indicating the name of a
 #'     logical assay of \code{sce} to use for filling the distribution plots.
 #' @param xlab Character scalar providing the x-axis label for the plot.
+#' @param plotType Character scalar indicating the type of plot to make
+#'     (either "histogram" or "density").
 #'
 #' @export
 #' @author Charlotte Soneson
@@ -30,13 +32,15 @@
 #' @importFrom rlang .data
 #'
 plotImputationDistribution <- function(sce, assayToPlot, assayImputation,
-                                       xlab = "") {
+                                       xlab = "", plotType = "histogram") {
     .assertVector(x = sce, type = "SummarizedExperiment")
     .assertScalar(x = assayToPlot, type = "character",
                   validValues = SummarizedExperiment::assayNames(sce))
     .assertScalar(x = assayImputation, type = "character",
                   validValues = SummarizedExperiment::assayNames(sce))
     .assertScalar(x = xlab, type = "character")
+    .assertScalar(x = plotType, type = "character",
+                  validValues = c("histogram", "density"))
 
     plotdf <- as.data.frame(
         SummarizedExperiment::assay(sce, assayToPlot)) %>%
@@ -49,11 +53,26 @@ plotImputationDistribution <- function(sce, assayToPlot, assayImputation,
                 tidyr::gather(key = "sample", value = "imputed", -"pid"),
             by = c("pid", "sample")
         )
-    ggplot2::ggplot(plotdf, ggplot2::aes(x = .data$log2intensity,
-                                         fill = .data$imputed)) +
-        ggplot2::geom_histogram(bins = 50) +
-        ggplot2::facet_wrap(~ sample) +
-        ggplot2::theme_bw() + ggplot2::labs(x = xlab) +
-        ggplot2::scale_fill_manual(values = c(`TRUE` = "grey",
-                                              `FALSE` = "firebrick1"))
+    if (plotType == "histogram") {
+        ggplot2::ggplot(plotdf, ggplot2::aes(x = .data$log2intensity,
+                                             fill = .data$imputed)) +
+            ggplot2::geom_histogram(bins = 50) +
+            ggplot2::facet_wrap(~ sample) +
+            ggplot2::theme_bw() + ggplot2::labs(x = xlab) +
+            ggplot2::scale_fill_manual(values = c(`TRUE` = "grey",
+                                                  `FALSE` = "firebrick1"))
+    } else if (plotType == "density") {
+        ggplot2::ggplot(plotdf, ggplot2::aes(x = .data$log2intensity,
+                                             color = .data$imputed)) +
+            ggplot2::geom_density(linewidth = 1.5) +
+            ggplot2::facet_wrap(~ sample) +
+            ggplot2::theme_bw() + ggplot2::labs(x = xlab) +
+            ggplot2::scale_color_manual(values = c(`TRUE` = "grey",
+                                                   `FALSE` = "firebrick1"))
+    } else {
+        ## Should never end up here as the parameter is checked above
+        #nocov start
+        stop("Unknown value of the plotType parameter")
+        #nocov end
+    }
 }